pyproject.toml
[project]
name = "yio-minions"
description = "A framework to develop external tasks for camunda process engine in python. Especially for yio ecosystem."
readme = "README.md"

version = "2.0.6"
authors = [
    { name = "yio gmbh", email = "hello@yio.at" }
]

dependencies = [
    "aetpiref>=0.1.6,<1.0.0",
    "aiolimiter>=1.2.1",
    "asyncer~=0.0.2",
    "culsans>=0.11.0",
    "email-validator<=2.2",
    "httpx>=0.24.1",
    "httpx-auth~=0.18",
    "jsonschema2md2>=1.0.0",
    "mkdocs-material~=9.4",
    "mkdocs-simple-hooks~=0.1.5",
    "pydantic>=2.0,<3.0",
    "pydantic-settings>=2.13.1",
    "pymdown-extensions~=10.5",
    "python-dotenv~=1.0",
    "python-magic~=0.4",
    "rich>=14.3.3",
    "structlog>=25.5.0",
    "structlog-config>=0.10.0",
    "tenacity>=9.1.4",
    "typer>=0.24.1,<0.25.0",
]
requires-python = ">=3.11"

keywords = [
    "minions",
    "yio",
    "process engine",
    "aetpi - async external task processing interfance"
]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]



[project.urls]
Homepage = "https://yio.at"
Documentation = "https://docs.yio.at/products/yio-minions"
Repository = "https://github.com/yiogmbh/fastbpmn.git"
Issues = "https://github.com/yiogmbh/fastbpmn/issues"
Changelog="https://github.com/yiogmbh/fastbpmn/blob/main/CHANGELOG.md"



# no telemetry yet
# [project.optional-dependencies]



[project.scripts]
readiness = "yio_minions.readiness:main"
squirrel = "squirrel.app:app"

[build-system]
requires = ["uv_build>=0.11.3,<0.12"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = ["yio_minions", "squirrel"]
namespace = true

[dependency-groups]
dev = [
    { include-group = "test" },
    { include-group = "lint"},
    { include-group = "tox"},
    { include-group = "b2v"},
    "prek>=0.3.8",
]
lint = [
    "ruff",
    "ty==0.0.27",
]
test = [
    "pytest>=8.1",
    "pytest-cov==7.0.0",
    "pytest-env==1.5.0",
    "pytest-httpx",
    "pyfakefs==6.1.1",
    "pytest-asyncio",
    "pytest-freezegun==0.4.2",
    "freezegun==1.5.5",
    "assertpy==1.1",
    "testcontainers[postgres]>=4.14.1",
    "polyfactory>=3.3.0",
]
# we use this dependency group to prepare the environment for running tests
tox = [
    "tox",
    "tox-uv",
    "keyring",
]
b2v = [
    "git-cliff"
]

[tool.uv]
keyring-provider = "disabled"

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple/"
default = true

[tool.ruff]
target-version = "py311"

[tool.ruff.lint.isort]
force-single-line = true


[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "module"
asyncio_default_test_loop_scope = "module"
addopts =  "--ignore tests"
